requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-serializer==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
python-decouple==3.8